Change Player::_pSplLvl to uint8_t

+++ Source/debug.cpp
@@ -92,7 +92,7 @@ uint32_t glMid2Seed[NUMLEVELS];
 uint32_t glMid3Seed[NUMLEVELS];
 uint32_t glEndSeed[NUMLEVELS];
 
-void SetSpellLevelCheat(SpellID spl, int spllvl)
+void SetSpellLevelCheat(SpellID spl, uint8_t spllvl)
 {
 	Player &myPlayer = *MyPlayer;
 
@@ -529,7 +529,7 @@ std::string DebugCmdMinStats(const string_view parameter)
 
 std::string DebugCmdSetSpellsLevel(const string_view parameter)
 {
-	int level = std::max(0, atoi(parameter.data()));
+	uint8_t level = static_cast<uint8_t>(std::max(0, atoi(parameter.data())));
 	for (int i = static_cast<int8_t>(SpellID::Firebolt); i < MAX_SPELLS; i++) {
 		if (GetSpellBookLevel(static_cast<SpellID>(i)) != -1) {
 			SetSpellLevelCheat(static_cast<SpellID>(i), level);
